fix: tailwind install
